refactor : userService 의 createUser 메서드를 joiningPool 로 이동
userService 의 createUser 메서드를 joiningPool 로 이동
