+++ waldreg/service-layer/user/src/main/java/org/waldreg/user/management/user/DefaultUserManager.java
@@ -5,7 +5,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.waldreg.character.exception.UnknownCharacterException;
 import org.waldreg.user.dto.UserDto;
-import org.waldreg.user.exception.DuplicatedUserIdException;
 import org.waldreg.user.exception.InvalidRangeException;
 import org.waldreg.user.exception.UnknownIdException;
 import org.waldreg.user.exception.UnknownUserIdException;
@@ -18,43 +17,12 @@ import org.waldreg.user.spi.CharacterRepository;
 public class DefaultUserManager implements UserManager{
 
     private final UserRepository userRepository;
-    private final JoiningPoolRepository joiningPoolRepository;
     private final CharacterRepository characterRepository;
 
     @Autowired
-    public DefaultUserManager(UserRepository userRepository, JoiningPoolRepository joiningPoolRepository, CharacterRepository characterRepository){
+    public DefaultUserManager(UserRepository userRepository, CharacterRepository characterRepository){
         this.userRepository = userRepository;
-        this.joiningPoolRepository = joiningPoolRepository;
         this.characterRepository = characterRepository;
-    }
-
-    @Override
-    public void createUser(UserDto userDto){
-        throwIfDuplicatedUserId(userDto.getUserId());
-        setDefaultCharacter(userDto);
-        joiningPoolRepository.createUser(userDto);
-    }
-
-    public void throwIfDuplicatedUserId(String userId){
-        if (userRepository.isExistUserId(userId) || joiningPoolRepository.isExistUserId(userId)){
-            throw new DuplicatedUserIdException("Duplicated user_id \"" + userId + "\"");
-        }
-    }
-
-    private void setDefaultCharacter(UserDto userDto){
-        if (userDto.getUserId().equals("Admin")){
-            throwIfUnknownCharacter("Admin");
-            userDto.setCharacter("Admin");
-            return;
-        }
-        throwIfUnknownCharacter("Guest");
-        userDto.setCharacter("Guest");
-    }
-
-    private void throwIfUnknownCharacter(String characterName){
-        if (!characterRepository.isExistCharacterName(characterName)){
-            throw new UnknownCharacterException(characterName);
-        }
     }
 
     @Override
