+++ waldreg/service-layer/user/src/main/java/org/waldreg/user/management/user/UserManager.java
@@ -5,8 +5,6 @@ import org.waldreg.user.dto.UserDto;
 
 public interface UserManager{
 
-    void createUser(UserDto userDto);
-
     UserDto readUserById(int id);
 
     UserDto readUserByUserId(String userId);
